refactor: useENhancedWithGas to be a regular enhancedWithGas fn
